type fix

+++ parsee/extraction/models/llm_models/replicate_model.py
@@ -56,5 +56,5 @@ class ReplicateModel(LLMBaseModel):
         final_prompt, num_tokens_input = truncate_prompt(prompt, self.encoding, self.max_tokens_question)
         response = self._call_api(final_prompt)
         tokens_response = len(get_tokens_encoded(response, self.encoding))
-        final_cost = ((tokens_response + num_tokens_input) * (self.model.price_per_1k_tokens/1000)) if self.model.price_per_1k_tokens is not None else Decimal(0)
+        final_cost = Decimal((tokens_response + num_tokens_input) * (self.model.price_per_1k_tokens/1000)) if self.model.price_per_1k_tokens is not None else Decimal(0)
         return response, final_cost
